test: :white_check_mark: Updated returnCurrencyFormat's return part from parseInt to parseFloat so that it doesn't ignore the decimal places. Also updated the returnEmailValidity's regex to accommodate a wider variation of valid and invalid email addresses. Completed unit tests for utility functions (convertArrayToDropdown, returnArrayToList, returnCurrencyFormat, returnDaysSince, returnEmailValidity).

+++ utils/returnCurrencyFormat.ts
@@ -7,5 +7,5 @@ export function returnCurrencyFormat(amount?: string | number, options?: Intl.Nu
 
   const mergedOptions: Intl.NumberFormatOptions = defu(options, { style: 'currency', currency: 'AUD' } as Intl.NumberFormatOptions)
 
-  return new Intl.NumberFormat("en-AU", mergedOptions).format(parseInt(amount.toString()));
+  return new Intl.NumberFormat("en-AU", mergedOptions).format(parseFloat(amount.toString()));
 }
